requirements.txt
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
python-ulid==3.0.0
SQLAlchemy==2.0.40
tsidpy==1.1.5
typing_extensions==4.13.2
Werkzeug==3.1.3
